Move auxiliary methods to the transform module
One of the goals of the refactor is to move out of the utils dir
